% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/largen_binom_ci.R
\name{largen_binom_ci}
\alias{largen_binom_ci}
\title{Large sample CI for a binomial probability}
\usage{
largen_binom_ci(x, n, conf = 0.95)
}
\arguments{
\item{x}{number of binomial "successes" (proportion numerator)}

\item{n}{number of binomial "trials" (proportion denominator)}

\item{conf}{Confidence coefficient for CIs as proportion: .95 for 95\% CIs is default}
}
\value{
Estimate of binomial probability (sample proportion), large-sample se, and CI
}
\description{
Large sample CI for a binomial probability
}
\examples{
largen_binom_ci(15, 44)
largen_binom_ci(0, 20)
}
