% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust.R
\name{robust}
\alias{robust}
\title{Coefficient summary table for linear models with robust standard errors and robust CIs}
\usage{
robust(model, type = c("HC1"), level = 0.95, digits = 2)
}
\arguments{
\item{model}{A model object created by the fit of linear model using lm()}

\item{type}{Character string giving type of robust standard error: HC1 is default}

\item{level}{Confidence coefficient for CIs as proportion: .95 for 95\% CIs is default}

\item{digits}{Number of digits to round coefficients, SEs and CIs to in output.  Default is 2.}
}
\value{
A table with coefficient estimates, and CIs and P-values based on robust standard errors
}
\description{
Coefficient summary table for linear models with robust standard errors and robust CIs
}
\examples{
model <- lm(mpg ~ gear, data = mtcars)
robust(model)
}
